Update logic to be recursive
